script to include macro cfg file

+++ scripts/ensure_include_cfgs.sh
@@ -2,14 +2,16 @@
 
 set -e
 
-KLIPPER_CONF_DIR="$1"
-PRINTER_CONF_FILE="${KLIPPER_CONF_DIR}/printer.cfg"
+PRINTER_CONF_FILE="$1"
+KLIPPER_CONF_DIR=$(dirname "${PRINTER_CONF_FILE}")
 
 OBICO_DIR=$(readlink -f $(dirname "$0"))/..
 MACRO_CFG="${OBICO_DIR}/include_cfgs/moonraker_obico_macros.cfg"
 
-ls "${MACRO_CFG}" > /dev/null # make sure file exists
-ln -sf "${MACRO_CFG}" "${KLIPPER_CONF_DIR}/moonraker_obico_macros.cfg"
+if [ ! -f ${KLIPPER_CONF_DIR}/moonraker_obico_macros.cfg ]; then
+  ls "${MACRO_CFG}" > /dev/null # make sure file exists
+  ln -sf "${MACRO_CFG}" "${KLIPPER_CONF_DIR}/moonraker_obico_macros.cfg"
+fi
 
 if ! grep -q "include moonraker_obico_macros.cfg" "${PRINTER_CONF_FILE}" ; then
   sed -i "1 i [include moonraker_obico_macros.cfg]" "${PRINTER_CONF_FILE}"
